libcollectdclient: Move type definitions into their own header.
The "types.h" header is included by "client.h", so existing code does not
need to be updated.
